--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,7 @@ module github.com/segmentio/kafka-go
 go 1.15
 
 require (
-	github.com/klauspost/compress v1.15.7
+	github.com/klauspost/compress v1.15.9
 	github.com/pierrec/lz4/v4 v4.1.15
 	github.com/stretchr/testify v1.8.0
 	github.com/xdg/scram v1.0.5