Before:
```
BenchmarkCompression/gzip/compress-32                 81          13573535 ns/op           8.77 MB/s       10115 B/op          1 allocs/op
BenchmarkCompression/gzip/decompress-32              580           2023225 ns/op         959.10 MB/s        7563 B/op         53 allocs/op

  gzip:         93.86%
```

After:
```
BenchmarkCompression/gzip/compress-32                224           5340019 ns/op          24.42 MB/s        4838 B/op          1 allocs/op
BenchmarkCompression/gzip/decompress-32              692           1713771 ns/op        1132.28 MB/s         566 B/op         34 allocs/op

  gzip:         93.28%
````

Pretty typical scenario for default settings, 2-3x faster, sometimes at a slight compression loss. This provides a much better "default" trade-off than the stdlib.

For people regretting the 0.6% loss the higher compression levels can be used.

Decompression typically has even bigger margin, but it depends on the input.